Edit rate argument on stores model
change integer to double

+++ lib/models/stores_model.dart
@@ -1,7 +1,7 @@
 class StoresModel {
   final String name;
   final String address;
-  final int rate;
+  final double rate;
   final String phone;
   final String reviews;
   StoresModel({
